changed from parseInt to parseFloat

+++ src/client-pages/account/AirdropTokenForm.tsx
@@ -17,7 +17,7 @@ export default function AirdropTokenForm({
 
   const submitHandler = async (e: FormEvent<HTMLFormElement>) => {
     e.preventDefault();
-    const value = parseInt(amountRef.current?.value || "asd");
+    const value = parseFloat(amountRef.current?.value || "asd");
 
     if (isNaN(value) || !wallet) {
       return;
